更新package.json版本至0.2.6，修改播放器的seek属性类型为Int64，并相应调整相关方法以支持新的时间格式。

+++ ios/ExpoVlcPlayerModule.swift
@@ -38,7 +38,7 @@ public class ExpoVlcPlayerModule: Module {
             }
 
             Prop("seek") {
-                (view: ExpoVlcPlayerView, seek: Float) in
+                (view: ExpoVlcPlayerView, seek: Int64) in
                 if view.seek != seek {
                     view.seek = seek
                 }
